site_mapper/requirements.txt
selenium==4.18.1
beautifulsoup4==4.12.3
requests==2.31.0
PyPDF2==3.0.1
pdfminer.six==20221105
python-docx==1.1.0
openpyxl==3.1.2
uuid==1.30